% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_multiple_circles.R
\name{add_multiple_circles}
\alias{add_multiple_circles}
\title{Add multiple concentric circles to a ggplot object}
\usage{
add_multiple_circles(
  radii = c(0.25, 0.5, 0.75),
  circle_color = "grey20",
  circle_alpha = 1,
  circle_size = 0.5
)
}
\arguments{
\item{radii}{A vector of radii for the concentric circles (default is c(0.25, 0.5, 0.75, 1))}

\item{circle_color}{The color of the circles (default is "grey40")}

\item{circle_alpha}{The transparency of the circles (default is 1)}

\item{circle_size}{The size of the circle lines (default is 1)}
}
\value{
A list of layers for concentric circles
}
\description{
This function creates a list of layers for concentric circles with specified radii that can be added to a ggplot object.
The function takes optional arguments to customize the appearance of the circles.
}
\examples{
# Create concentric circles with custom appearance
circles <- add_multiple_circles(radii = c(0.2, 0.4, 0.6, 0.8, 1),
                                circle_color = "grey",
                                circle_alpha = 0.8,
                                circle_size = 1)

# Create a ggplot object
library(ggplot2)
p <- ggplot() + coord_fixed()

# Add the concentric circles
p_modified <- p + circles

}
