% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibrate_positions}
\alias{calibrate_positions}
\alias{calibrate_positions,TrajSet,CalModel-method}
\title{Calibrate TrajSet positions using a camera model}
\usage{
calibrate_positions(x, model)

\S4method{calibrate_positions}{TrajSet,CalModel}(x, model)
}
\arguments{
\item{x}{TrajSet with `x`/`y` coordinates}

\item{model}{`CalModel` object containing calibration parameters}
}
\value{
TrajSet with updated `x`/`y` and angles in metric space
}
\description{
Applies camera intrinsics + radial distortion to convert pixel coordinates to metric space.
}
\seealso{
Other calibration: 
\code{\link{CalModel-class}}
}
\concept{calibration}
