% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_switch_axes.R
\name{calibration_switch_axes}
\alias{calibration_switch_axes}
\title{Switch axes of the given intrinsic camera matrix}
\usage{
calibration_switch_axes(K)
}
\arguments{
\item{K}{intrinsic camera matrix}
}
\value{
modified intrinsic camera matrix
}
\description{
Switch axes of the given intrinsic camera matrix
}
\keyword{internal}
