% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\docType{class}
\name{CalModel-class}
\alias{CalModel-class}
\title{Camera calibration model for trajectory correction}
\description{
Camera calibration model for trajectory correction
}
\section{Slots}{

\describe{
\item{\code{K}}{3x3 intrinsic calibration matrix}

\item{\code{k}}{Numeric vector or list of distortion coefficients (radial first, optional tangential)}

\item{\code{F}}{Numeric scalar or length-2 vector giving focal length scaling (e.g., mm per pixel)}
}}

\seealso{
\code{\link{calibrate_positions}}

Other calibration:
\code{\link[=cal_model]{cal_model()}},
\code{\link[=calibrate_positions]{calibrate_positions()}},
\code{\link[=read_calibration]{read_calibration()}}
}
\concept{calibration}
