% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{cam_cal_pt}
\alias{cam_cal_pt}
\title{Calibrate the camera on the given xy coordinate}
\usage{
cam_cal_pt(x, y, K, k, F, max_iterations = 5, tolerance = 1e-09)
}
\arguments{
\item{x}{x-coordinate in pixels}

\item{y}{y-coordinate in pixels}

\item{K}{intrinsic camera matrix}

\item{k}{distortion coefficients (see [radial_distort()] for supported layouts)}

\item{F}{focal length in millimeters (scalar or length-2 vector)}

\item{max_iterations}{Maximum number of iterations passed to [radial_distort()].}

\item{tolerance}{Convergence tolerance passed to [radial_distort()].}
}
\value{
calibrated point
}
\description{
Calibrate the camera on the given xy coordinate
}
\examples{
K <- matrix(c(784.948340421183, 0, 0,
              0, 782.554388639436, 0,
              939.047051578744, 528.896744808718, 1), ncol = 3, byrow = TRUE)
k <- c(-0.289927776375773, 0.0392224238600441)
F <- rep(680, 2)
cam_cal_pt(100, 100, K, k, F)
}
