% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_mean.R
\name{circular_mean}
\alias{circular_mean}
\title{Calculate the circular mean of a set of angles in radians}
\usage{
circular_mean(angles, normalize_range = TRUE)
}
\arguments{
\item{angles}{A numeric vector of angles in radians}

\item{normalize_range}{Logical, if TRUE (default), the output will be in the range (-pi, pi], otherwise in the range [0, 2 * pi)}
}
\value{
The circular mean angle in radians
}
\description{
Calculate the circular mean of a set of angles in radians
}
\examples{
angles <- c(0, pi, pi/2, 3*pi/2)
circular_mean(angles)
circular_mean(angles, normalize_range = FALSE)
}
