% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_stdev.R
\name{circular_sd_from_R}
\alias{circular_sd_from_R}
\title{Calculate the circular standard deviation using the mean resultant length (R)}
\usage{
circular_sd_from_R(R)
}
\arguments{
\item{R}{The mean resultant length (R)}
}
\value{
The circular standard deviation
}
\description{
Calculate the circular standard deviation using the mean resultant length (R)
}
\examples{
R <- mean_resultant_length(c(0, pi, pi/2, 3*pi/2))
circular_sd_from_R(R)
}
