% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plotting.R
\name{directedness_arrow}
\alias{directedness_arrow}
\title{Make mean resultant length arrow}
\usage{
directedness_arrow(
  data,
  angle_col,
  arrow_head_cm = 0.2,
  colour = "gray",
  size = 2
)
}
\arguments{
\item{data}{Data frame containing the angle column.}

\item{angle_col}{Column containing angles in radians.}

\item{arrow_head_cm}{Length of the arrowhead in centimetres.}

\item{colour}{Colour of the arrow.}

\item{size}{Width of the arrow segment (applied to the geom's `linewidth`).}
}
\value{
A `geom_segment()` layer.
}
\description{
Computes the circular mean direction and resultant length, returning a
`geom_segment()` layer that can be added to an existing ggplot.
}
