% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headings.R
\name{.heading_registry}
\alias{.heading_registry}
\title{Build a data frame of arrow segments representing mean direction vectors
Length equals resultant_R; angle equals mean_dir}
\usage{
.heading_registry
}
\description{
Build a data frame of arrow segments representing mean direction vectors
Length equals resultant_R; angle equals mean_dir
}
