% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_tracks.R
\name{draw_tracks}
\alias{draw_tracks}
\title{Create geom_point or geom_line object for Cartesian coordinates}
\usage{
draw_tracks(data, geom = NULL, ...)
}
\arguments{
\item{data}{data}

\item{geom}{A character specifying the type of object to create, either "point" or "line" (default is "point")}

\item{...}{Additional arguments to be passed to the internal geom_point or geom_line functions}
}
\value{
A ggplot2 geom_point or geom_path object based on the input type
}
\description{
This function takes sets of Cartesian coordinates, and creates a ggplot2 object using
either geom_point or geom_line, based on the specified input parameter `type`.
}
\examples{
data <- data.frame(lon = c(1, 2, 3, 2, 0), lat = c(3, 4, 5, 7, 8),
cohort = c("A", "A", "B", "B", "C"))
geom_obj <- draw_tracks(data, mapping = aes(color = cohort))
ggplot(data, aes(x = lon, y = lat)) + geom_obj

}
