% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{focalize}
\alias{focalize}
\title{Normalize the given point by the focal length}
\usage{
focalize(xy, f)
}
\arguments{
\item{xy}{point}

\item{f}{focal length}
}
\value{
normalized point
}
\description{
Normalize the given point by the focal length
}
\keyword{internal}
