% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{frame_rate}
\alias{frame_rate}
\alias{frame_rate,Tracks-method}
\alias{set_frame_rate}
\alias{set_frame_rate,Tracks-method}
\title{Frame rate of a Tracks object}
\usage{
frame_rate(x)

\S4method{frame_rate}{Tracks}(x)

set_frame_rate(x, fps)

\S4method{set_frame_rate}{Tracks}(x, fps)
}
\arguments{
\item{x}{A [Tracks] object.}

\item{fps}{A single positive number, frames per second.}
}
\value{
`frame_rate()` the stored fps (or `NULL`); `set_frame_rate()` the
  modified `Tracks`.
}
\description{
Attach a capture frame rate (frames per second) to a [Tracks] so the time
aspect of frame-indexed tracks can be represented in real seconds. Stored in
the object's metadata; the time/frame column is not altered.
}
\seealso{
[elapsed_seconds()], [track_duration()]
}
