% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tracked_object_pos.R
\name{get_tracked_object_pos}
\alias{get_tracked_object_pos}
\title{Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.}
\usage{
get_tracked_object_pos(
  trial_limits,
  animal_track,
  circ0 = NULL,
  circ1 = NULL,
  radius_criterion = NULL
)
}
\arguments{
\item{trial_limits}{A data frame containing a row for each tracked point.}

\item{animal_track}{The number of columns used by grid_wrap.}
}
\description{
Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.
}
\examples{
utils::data(urchin_tracks)
radiate(urchin_tracks, ncols=3, group1 = "arc", group2 = "id")
}
