% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_limits.R
\name{get_trial_limits}
\alias{get_trial_limits}
\title{Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.}
\usage{
get_trial_limits(landmarks, animal_track, file_tbl, vid_num, midpoint = NULL)
}
\arguments{
\item{landmarks}{A data frame containing a row for each tracked point.}

\item{animal_track}{The number of columns used by grid_wrap.}

\item{file_tbl}{Inner circle drawn.}

\item{vid_num}{Outer circle drawn.}
}
\value{
A .
}
\description{
Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.
}
\examples{
utils::data(urchin_tracks)
radiate(urchin_tracks, ncols=3, group1 = "arc", group2 = "id")
}
