% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headings.R
\name{gg_add_circ_mean}
\alias{gg_add_circ_mean}
\title{Add mean direction arrows to a ggplot polar plot}
\usage{
gg_add_circ_mean(
  p,
  segments_df,
  color = "black",
  linewidth = 0.8,
  arrow_spec = ggplot2::arrow(length = grid::unit(0.02, "npc")),
  inherit_aes = FALSE
)
}
\arguments{
\item{p}{ggplot built from gg_traj(...) or your own polar builder}

\item{segments_df}{data from circ_mean_segments()}

\item{color}{arrow color}

\item{linewidth}{segment linewidth}

\item{arrow_spec}{ggplot2::arrow(...) for arrow heads}

\item{inherit_aes}{whether to inherit aes (usually FALSE)}
}
\description{
Add mean direction arrows to a ggplot polar plot
}
