% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headings.R
\name{list_heading_rules}
\alias{list_heading_rules}
\title{List registered custom heading rules}
\usage{
list_heading_rules()
}
\value{
Sorted character vector of registered rule names.
}
\description{
List registered custom heading rules
}
\seealso{
\code{\link{register_heading_rule}}
}
