% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tracks2.R
\name{load_tracks2}
\alias{load_tracks2}
\title{Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.}
\usage{
load_tracks2(file_tbl, df, track_dir, colnames)
}
\arguments{
\item{file_tbl}{.}

\item{df}{.}

\item{track_dir}{The folder with the tracks.}

\item{colnames}{A named list of column names to replace the hardcoded ones.}
}
\description{
Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.
}
\examples{
utils::data(urchin_tracks)
radiate(urchin_tracks, ncols=3, colnames = list("arc" = "new_arc", "id" = "new_id"))
}
