% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_resultant_length.R
\name{mean_resultant_length}
\alias{mean_resultant_length}
\title{Calculate the mean resultant length of a set of angles in radians}
\usage{
mean_resultant_length(angles)
}
\arguments{
\item{angles}{A numeric vector of angles in radians}
}
\value{
The mean resultant length (R)
}
\description{
Calculate the mean resultant length of a set of angles in radians
}
\examples{
angles <- c(0, pi, pi/2, 3*pi/2)
mean_resultant_length(angles)
}
