% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optically_center.R
\name{optically_center}
\alias{optically_center}
\title{Optically center the given point}
\usage{
optically_center(x, y, c)
}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{c}{optical center}
}
\value{
normalized point
}
\description{
Optically center the given point
}
\keyword{internal}
