% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_mapping.R
\name{rad_shepherd_clock}
\alias{rad_shepherd_clock}
\title{Wrap angles to the interval [0, 2*pi)}
\usage{
rad_shepherd_clock(theta)
}
\arguments{
\item{theta}{Numeric vector of angles (radians).}
}
\value{
Angles wrapped to [0, 2*pi).
}
\description{
Wrap angles to the interval [0, 2*pi)
}
\examples{
theta <- seq(from = -2, to = 8, length.out = 6)
rad_shepherd_clock(theta)
}
