% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_unclock.R
\name{rad_unclock}
\alias{rad_unclock}
\title{Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.}
\usage{
rad_unclock(theta)
}
\arguments{
\item{theta}{Angle angle in radians in 'clockwise' format.}
}
\value{
An angle in radians in a conventional unit circle format.
}
\description{
Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.
}
\examples{
a <- seq(from=0,to=2*pi,length.out=6)
mapply(rad_unclock,a)
}
