% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration_capture.R
\name{read_calibration_points}
\alias{read_calibration_points}
\title{Read calibration point CSV}
\usage{
read_calibration_points(path)
}
\arguments{
\item{path}{File path previously written by
[write_calibration_points()].}
}
\value{
List of 2-column matrices containing image points per frame.
}
\description{
Read calibration point CSV
}
