% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{read_tracks_format}
\alias{read_tracks_format}
\title{Construct a Tracks from a *format* spec (registered name or inline list)}
\usage{
read_tracks_format(x, format, ...)
}
\arguments{
\item{x}{data.frame or path(s)}

\item{format}{registered format name or list spec}

\item{...}{extra args override spec fields}
}
\description{
Construct a Tracks from a *format* spec (registered name or inline list)
}
