% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{register_loader_dialect}
\alias{register_loader_dialect}
\title{Register a custom loader dialect
The function must accept (x, ...) and return a data.frame in long form with columns at least id,time and one of (angle) or (x,y)}
\usage{
register_loader_dialect(name, fun, overwrite = FALSE)
}
\arguments{
\item{name}{Unique dialect name}

\item{fun}{Function that accepts (x, ...) and returns a data.frame with id/time/angle or id/time/x/y}

\item{overwrite}{Replace an existing dialect registered with the same name}
}
\description{
Register a custom loader dialect
The function must accept (x, ...) and return a data.frame in long form with columns at least id,time and one of (angle) or (x,y)
}
