% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimeter_labs.R
\name{scale_months}
\alias{scale_months}
\title{Perimeter scale: months of the year}
\usage{
scale_months(format = c("abbr", "initial", "number"))
}
\arguments{
\item{format}{`"abbr"` (Jan...Dec, default), `"initial"` (J, F, M, ...), or
`"number"` (1...12).}
}
\value{
A perimeter-scale list for [perimeter_labs()].
}
\description{
Perimeter scale: months of the year
}
\examples{
scale_months("initial")
}
\seealso{
[perimeter_labs()], [scale_cardinal()], [scale_clock()],
  [scale_seconds()]
}
