% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{scaled_xy2mm}
\alias{scaled_xy2mm}
\title{Convert the given point from pixels to millimeters}
\usage{
scaled_xy2mm(xy, f, F)
}
\arguments{
\item{xy}{point}

\item{f}{focal length in pixels}

\item{F}{focal length in millimeters}
}
\value{
point in millimeters
}
\description{
Convert the given point from pixels to millimeters
}
\keyword{internal}
