% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaders.R
\name{TrajSet_read_format}
\alias{TrajSet_read_format}
\title{Construct a TrajSet from a *format* spec (registered name or inline list)}
\usage{
TrajSet_read_format(x, format, ...)
}
\arguments{
\item{x}{data.frame or path(s)}

\item{format}{registered format name or list spec}

\item{...}{extra args override spec fields}
}
\description{
Construct a TrajSet from a *format* spec (registered name or inline list)
}
