% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ticks.R
\name{add_ticks}
\alias{add_ticks}
\title{Create annotations for degrees at diagonals}
\usage{
add_ticks()
}
\arguments{
\item{degrees}{A boolean indicating if degree marks should be created (default is TRUE)}
}
\value{
A geom_segment object of ticks
}
\description{
This function creates a list of annotation layers for degree marks at the diagonals.
The function takes one argument, a boolean value to specify if degree marks should be created or not.
}
\examples{
# Create degree annotations
ticks <- add_ticks(TRUE)

# Create a ggplot object
library(ggplot2)
p <- ggplot() + coord_polar()

# Add degree marks at diagonals
p_modified <- p + ticks

}
