% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plotting.R
\name{add_ticks}
\alias{add_ticks}
\title{Create tick marks at the cardinal directions.}
\usage{
add_ticks()
}
\value{
A `geom_segment()` layer.
}
\description{
Generates a `geom_segment()` layer containing small tick marks at north,
south, east, and west. The layer can be added to any ggplot.
}
\examples{
library(ggplot2)
ggplot() +
  coord_fixed() +
  add_ticks()
}
