% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_cal_many.R
\name{cam_cal_many}
\alias{cam_cal_many}
\title{Calibrate the camera on the given set of points}
\usage{
cam_cal_many(points_idx, K, k, F)
}
\arguments{
\item{points_idx}{matrix of points}

\item{K}{intrinsic camera matrix}

\item{k}{radial distortion coefficients}

\item{F}{focal length in millimeters}
}
\value{
matrix of calibrated points
}
\description{
Calibrate the camera on the given set of points
}
\examples{
K <- matrix(c(784.948340421183, 0, 0,
              0, 782.554388639436, 0,
              939.047051578744, 528.896744808718, 1), ncol = 3, byrow = TRUE)
k <- c(-0.289927776375773, 0.0392224238600441)
F <- rep(680, 2)
points <- matrix(c(100, 100, 200, 200, 300, 300), ncol = 2, byrow = TRUE)
cam_cal_many(points, K, k, F)
}
