% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cam_cal_pt.R
\name{cam_cal_pt}
\alias{cam_cal_pt}
\title{Calibrate the camera on the given xy coordinate}
\usage{
cam_cal_pt(x, y, K, k, F)
}
\arguments{
\item{x}{x-coordinate}

\item{y}{y-coordinate}

\item{K}{intrinsic camera matrix}

\item{k}{radial distortion coefficients}

\item{F}{focal length in millimeters}
}
\value{
calibrated point
}
\description{
Calibrate the camera on the given xy coordinate
}
\examples{
K <- matrix(c(784.948340421183, 0, 0,
              0, 782.554388639436, 0,
              939.047051578744, 528.896744808718, 1), ncol = 3, byrow = TRUE)
k <- c(-0.289927776375773, 0.0392224238600441)
F <- rep(680, 2)
cam_cal_pt(100, 100, K, k, F)
}
