% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajSet.R, R/circular_statistics.R
\name{circ_summary}
\alias{circ_summary}
\alias{circ_summary,TrajSet-method}
\title{Circular summaries per trajectory}
\usage{
circ_summary(x, w = NULL, by = c("id", "global"))

circ_summary(x, w = NULL, by = c("id", "global"))

\S4method{circ_summary}{TrajSet}(x, w = NULL, by = c("id", "global"))
}
\arguments{
\item{x}{TrajSet input}

\item{w}{Optional weight column name}

\item{by}{Summarise per id or globally across all observations}
}
\value{
data.frame(id, n, t_start, t_end, mean_dir, resultant_R, kappa)

data.frame with summary metrics per group
}
\description{
Circular summaries per trajectory

Circular summary statistics for trajectories
}
