% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_labs.R
\name{degree_labs}
\alias{degree_labs}
\title{Create annotations for degrees at diagonals}
\usage{
degree_labs()
}
\arguments{
\item{degrees}{A boolean indicating if degree marks should be created (default is TRUE)}
}
\value{
A list of ggplot2 annotation layers for degree marks at the diagonals if `degrees` is TRUE
}
\description{
This function creates a list of annotation layers for degree marks at the diagonals.
The function takes one argument, a boolean value to specify if degree marks should be created or not.
}
\examples{
# Create degree annotations
degree_labs <- degree_labs(TRUE)

# Create a ggplot object
library(ggplot2)
p <- ggplot() + coord_polar()

# Add degree marks at diagonals
p_modified <- p + degree_labs

}
