% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plotting.R
\name{degree_labs}
\alias{degree_labs}
\title{Label the four diagonal directions.}
\usage{
degree_labs()
}
\value{
A list of ggplot2 annotation layers.
}
\description{
Provides a list of annotation layers that mark 45, 135, 225, and 315 degrees on a
unit circle.
}
\examples{
library(ggplot2)
ggplot() +
  coord_fixed() +
  degree_labs()
}
