% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directedness_arrow.R
\name{directedness_arrow}
\alias{directedness_arrow}
\title{Make mean resultant length arrow}
\usage{
directedness_arrow(dataframe, arrow_head_cm = NULL, color = NULL, size = NULL)
}
\arguments{
\item{arrow_head_cm}{The suffix of files containing landmark coordinates.}

\item{color}{The suffix of files containing track coordinates.}

\item{size}{Size of the segment.}

\item{trials_df}{The directory in which to look for the landmark files. Defaults to the current working directory.}
}
\value{
A data frame of file names.
#examples
#import_tracks(dir=data)
}
\description{
Make mean resultant length arrow
}
