% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headings.R
\name{.heading_registry}
\alias{.heading_registry}
\title{Register a custom heading rule by name.
The function must accept (df, cols, ...) and return a data.frame with columns id, time, heading (radians)}
\usage{
.heading_registry
}
\description{
Register a custom heading rule by name.
The function must accept (df, cols, ...) and return a data.frame with columns id, time, heading (radians)
}
