% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plotting.R
\name{draw_tracks}
\alias{draw_tracks}
\title{Create geom layers for Cartesian track coordinates}
\usage{
draw_tracks(data, x_col, y_col, geom = "path", mapping = NULL, ...)
}
\arguments{
\item{data}{Data frame that will be plotted.}

\item{x_col}{Name of the column mapped to the x aesthetic.}

\item{y_col}{Name of the column mapped to the y aesthetic.}

\item{geom}{Either a character vector (`"path"`/`"point"`) or a ggplot2 geom
function (e.g. [ggplot2::geom_path]).}

\item{mapping}{Optional aesthetics created with [ggplot2::aes()].}

\item{...}{Additional arguments passed on to the geom function.}
}
\value{
A list containing a single ggplot2 layer.
}
\description{
Create geom layers for Cartesian track coordinates
}
