% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_object_pos.R
\name{get_all_object_pos}
\alias{get_all_object_pos}
\title{Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.}
\usage{
get_all_object_pos(landmarks, animal_track, file_tbl, track_dir)
}
\arguments{
\item{landmarks}{.}

\item{animal_track}{.}

\item{file_tbl}{.}
}
\description{
Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.
}
\examples{
utils::data(urchin_tracks)
radiate(urchin_tracks, ncols=3, group1 = "arc", group2 = "id")
}
