% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tracks.R
\name{ids}
\alias{ids}
\alias{ids,Tracks-method}
\title{Trajectory identifiers of a Tracks}
\usage{
ids(x)

\S4method{ids}{Tracks}(x)
}
\arguments{
\item{x}{A [Tracks-class] object.}
}
\value{
The unique trajectory ids, in first-appearance order.
}
\description{
Trajectory identifiers of a Tracks
}
\examples{
ids(cpunctatus)
}
