% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_info.R
\name{import_info}
\alias{import_info}
\title{Import landmark coordinates from text files}
\usage{
import_info(filename, ...)
}
\arguments{
\item{filename}{The directory in which to look for the landmark files. Defaults to the current working directory.}
}
\value{
A dataframe of file names.
#examples
#import_tracks(dir=data)
}
\description{
Import landmark coordinates from text files
}
