% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_tracks.R
\name{import_tracks}
\alias{import_tracks}
\title{Import landmark coordinates from text files}
\usage{
import_tracks(dir, landmark_suffix = NULL, track_suffix = NULL)
}
\arguments{
\item{dir}{The directory in which to look for the landmark files. Defaults to the current working directory.}

\item{landmark_suffix}{The suffix of files containing landmark coordinates.}

\item{track_suffix}{The suffix of files containing track coordinates.}
}
\value{
A dataframe of file names.
#examples
#import_tracks(dir=data)
}
\description{
Import landmark coordinates from text files
}
