% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plotting.R
\name{line_circle_intercept}
\alias{line_circle_intercept}
\title{Find the intercept of a line with the unit circle}
\usage{
line_circle_intercept(x0, y0, x1, y1)
}
\arguments{
\item{x0, y0}{Starting coordinates of the vector.}

\item{x1, y1}{Ending coordinates of the vector.}
}
\value{
A tibble with the intersection point closest to `(x1, y1)`.
}
\description{
Find the intercept of a line with the unit circle
}
