% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_mapping.R
\name{rad2clock}
\alias{rad2clock}
\title{Convert unit-circle angles to clock orientation}
\usage{
rad2clock(theta)
}
\arguments{
\item{theta}{Angle in radians using the standard unit-circle convention.}
}
\value{
Angle in radians measured clockwise with zero at the top.
}
\description{
Convert unit-circle angles to clock orientation
}
\examples{
theta <- seq(from = pi/2, to = -pi/2, length.out = 5)
rad2clock(theta)
}
