% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2clock.R
\name{rad2clock}
\alias{rad2clock}
\title{Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.}
\usage{
rad2clock(theta)
}
\arguments{
\item{theta}{Angle angle in radians in conventional unit circle format.}
}
\value{
An angle in radians in a 'clockwise' format.
}
\description{
Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.
}
\examples{
a <- seq(from=pi/2,to=-pi/2,length.out=6)
mapply(rad2clock,a)
}
