% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_mapping.R
\name{rad_shepherd}
\alias{rad_shepherd}
\title{Wrap angles to the interval (-pi, pi]}
\usage{
rad_shepherd(theta)
}
\arguments{
\item{theta}{Numeric vector of angles (radians).}
}
\value{
Angles wrapped to (-pi, pi].
}
\description{
Wrap angles to the interval (-pi, pi]
}
\examples{
theta <- seq(from = -5, to = 5, length.out = 6)
rad_shepherd(theta)
}
