% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_shepherd.R
\name{rad_shepherd}
\alias{rad_shepherd}
\title{Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.}
\usage{
rad_shepherd(theta)
}
\arguments{
\item{theta}{Angle angle in radians in unit circle format, going anticlockwise.}
}
\value{
An angle in the same format, brought into range -pi,+pi
}
\description{
Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.
}
\examples{
a <- seq(from=-5,to=5,length.out=6)
mapply(rad_shepherd,a)
}
