% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plotting.R
\name{rad_shepherd_clock}
\alias{rad_shepherd_clock}
\title{Wrap clockwise angles to [0, 2*pi)}
\usage{
rad_shepherd_clock(theta)
}
\arguments{
\item{theta}{Angles in radians measured clockwise (clock format).}
}
\value{
Angles wrapped to [0, 2*pi).
}
\description{
Wrap clockwise angles to [0, 2*pi)
}
\examples{
theta <- seq(from = -2, to = 8, length.out = 6)
rad_shepherd_clock(theta)
}
