% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad_shepherd_clock.R
\name{rad_shepherd_clock}
\alias{rad_shepherd_clock}
\title{Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.}
\usage{
rad_shepherd_clock(theta)
}
\arguments{
\item{theta}{Angle angle in radians in 'clock' format, going clockwise}
}
\value{
An angle in the same format, brought into range -pi,+pi
}
\description{
Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.
}
\examples{
a <- seq(from=-2,to=8,length.out=6)
mapply(rad_shepherd_clock,a)
}
