% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plotting.R
\name{rad_unclock}
\alias{rad_unclock}
\title{Convert clock-oriented angles back to unit-circle orientation}
\usage{
rad_unclock(theta)
}
\arguments{
\item{theta}{Angle (radians) measured clockwise with zero at the top.}
}
\value{
Angle in radians using the standard unit-circle convention.
}
\description{
Convert clock-oriented angles back to unit-circle orientation
}
\examples{
theta <- seq(from = 0, to = 2 * pi, length.out = 5)
rad_unclock(theta)
}
