% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{radial_distort}
\alias{radial_distort}
\title{Apply radial distortion to the given point}
\usage{
radial_distort(xy, k)
}
\arguments{
\item{xy}{point}

\item{k}{radial distortion coefficients}
}
\value{
distorted point
}
\description{
Apply radial distortion to the given point
}
