% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiate.R
\name{radiate}
\alias{radiate}
\title{Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.}
\usage{
radiate(
  data,
  xcol,
  ycol,
  geom = NULL,
  group2 = NULL,
  circ1 = NULL,
  circ2 = NULL,
  ticks = NULL,
  degrees = NULL,
  legend = NULL,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  axes = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame containing a row for each tracked point.}

\item{group2}{Grouping variable within data used within panel.}

\item{ticks}{Logical indicating whether ticks are drawn at perimeter.}

\item{degrees}{Logical indicating whether to mark degrees at panel diagonals. Defaults to TRUE.}

\item{legend}{Logical indicating whether to show legend. Defaults to FALSE.}

\item{title}{Plot title}

\item{xlab}{Text to include as plot x label. Defaults to none.}

\item{ylab}{Text to include as plot y label. Defaults to none.}

\item{axes}{Logical indicating whether panel axes should be shown.}

\item{x_col}{A character specifying the name of the column representing the x coordinates}

\item{y_col}{A character specifying the name of the column representing the y coordinates}
}
\value{
A ggplot2 graphical object.
}
\description{
Make ggplot object of tracks radiating from circle centre. Accepts a data frame with x and y coordinates of the points, and optional grouping variables.
}
\examples{
utils::data(urchin_tracks)
radiate(urchin_tracks, ncols = 3, group1 = "arc", group2 = "id")
}
