% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiate.R
\name{radiate}
\alias{radiate}
\title{Make ggplot object of tracks radiating from circle centre.}
\usage{
radiate(
  data,
  ncols = NULL,
  circ1 = NULL,
  circ2 = NULL,
  group1 = NULL,
  group2 = NULL,
  ticks = NULL,
  degrees = NULL,
  legend = NULL,
  xlab = NULL,
  ylab = NULL
)
}
\arguments{
\item{data}{A data frame containing a row for each tracked point.}

\item{ncols}{The number of columns used by grid_wrap.}

\item{circ1}{Inner circle drawn.}

\item{circ2}{Outer circle drawn.}

\item{group1}{Grouping variable within data used by grid facet.}

\item{group2}{Grouping variable within data used within panel.}

\item{ticks}{Logical indicating whether ticks are drawn at perimeter.}

\item{degrees}{Logical indicating whether to mark degrees at panel diagonals. Defaults to TRUE.}

\item{legend}{Logical indicating whether to show legend. Defaults to FALSE.}

\item{xlab}{Text to include as plot x label. Defaults to none.}

\item{ylab}{Text to include as plot y label. Defaults to none.}
}
\value{
A ggplot2 graphical object.
}
\description{
Make ggplot object of tracks radiating from circle centre.
}
\examples{
utils::data(urchin_tracks)
radiate(urchin_tracks, ncols=3, group1 = "arc")
}
