% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radiatR-package.R
\docType{package}
\name{radiatR-package}
\alias{radiatR}
\alias{radiatR-package}
\title{radiatR: Plot Tracks in Circular Space}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
