% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circumference_labs.R
\name{scale_cardinal}
\alias{scale_cardinal}
\title{Circumference scale: cardinal compass directions}
\usage{
scale_cardinal(points = 4)
}
\arguments{
\item{points}{`4` (N/E/S/W, default) or `8` (adds the intercardinals
NE/SE/SW/NW).}
}
\value{
A circumference-scale list for [circumference_labs()].
}
\description{
Circumference scale: cardinal compass directions
}
\examples{
scale_cardinal(points = 8)
}
\seealso{
[circumference_labs()], [scale_clock()], [scale_months()],
  [scale_seconds()]
}
