% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_theme.R
\name{sparse_theme}
\alias{sparse_theme}
\title{Make mean resultant length arrow}
\usage{
sparse_theme(theme = NULL)
}
\arguments{
\item{theme}{The directory in which to look for the landmark files. Defaults to the current working directory.}
}
\value{
A theme object.
#examples
#import_tracks(dir=data)
}
\description{
Make mean resultant length arrow
}
