% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_plotting.R
\name{spartan_theme}
\alias{spartan_theme}
\title{Minimalist theme for radial track plots.}
\usage{
spartan_theme(...)
}
\arguments{
\item{...}{Additional arguments passed to [ggplot2::theme_grey()].}
}
\value{
A ggplot2 theme object.
}
\description{
Wrapper around [ggplot2::theme_grey()] that strips panel adornments to
emphasise the radial geometry used by the package.
}
