% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_roxygen2.R
\name{urchin_tracks}
\alias{urchin_tracks}
\title{This is data to be included in my package}
\format{
A data frame with 134339 rows and 23 variables
}
\usage{
data(urchin_tracks)
}
\description{
This is data to be included in my package
}
\references{
\url{jkirwan.org}
}
\author{
John D. Kirwan \email{john@jkirwan.org}
}
